reACT-agent/reACT-agent-v2/pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.3",
    "langchain-anthropic>=1.3.1",
    "langchain-google-genai>=4.1.3",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
